Reduce code duplication.
